Some prep work to support epub v3.
Pull the current writing code out into a separate file, since it's V2-specific. Stick it in a function that actually says it's V2.

Add a version number to the epub struct. Default it to v2, since that's what we support.

Add some version get/set code, though it'll yell if you try anything but v2 since that's all we support.